Change variables in Makfile for POSIX-compliance
POSIX only specifies one variable attribution:

    CC = cc

All the other (:=, ?=, ...) are implementation specific and interfere
with the .POSIX: target. For example, using

    CC ?= cc

Would make the $(CC) macro be assigned to 'cc' if the $CC environment
variable is empty.
The problem here is that the .POSIX target implies that if the $(CC)
macro is empty, it should be set to it's default, POSIX value: 'c99'.
Thus, the ?= assignment is useless in the caase of a POSIX makefile.
To use the environment variables, POSIX specifies the `-e` flags, which
will make make(1) use environment variables instead of it's own local
variables.

Please use it instead starting from now. Thanks.

Also, I'm sorry I asked dcat to change the CC declaration in the
previous commit. That's my fault bro, I didn't tested.