new tool: slw
This tool grabs the cursor and wait for a button press/release event.
When the event is received, the ID of the window that received the event
will be printed to stdout.

This is useful to select a random window interactively using the mouse
pointer.

Note: this adds a new "dependency" on the xcb-utils cursor library.